check if OpenInEditor-Lite is already installed

+++ macsetup.sh
@@ -1292,15 +1292,17 @@ fi
 #
 # https://github.com/Ji4n1ng/OpenInTerminal
 #
-installcask openineditor-lite
-defaults write wang.jianing.app.OpenInEditor-Lite LiteDefaultEditor "Visual Studio Code"
-open /Applications
-notification "drag openineditor-lite in Finder toolbar while pressing Command"
-curl -L https://github.com/Ji4n1ng/OpenInTerminal/releases/download/v1.2.0/Icons.zip  --output icons.zip
-unzip icons.zip
-rm icons.zip
-seticon icons/icon_vscode_dark.icns /Applications/OpenInEditor-Lite.app
-rm -rf icons
+if [[ ! $(isAppInstalled OpenInEditor-Lite) ]]; then
+    installcask openineditor-lite
+    defaults write wang.jianing.app.OpenInEditor-Lite LiteDefaultEditor "Visual Studio Code"
+    open /Applications
+    notification "drag openineditor-lite in Finder toolbar while pressing Command"
+    curl -L https://github.com/Ji4n1ng/OpenInTerminal/releases/download/v1.2.0/Icons.zip  --output icons.zip
+    unzip icons.zip
+    rm icons.zip
+    seticon icons/icon_vscode_dark.icns /Applications/OpenInEditor-Lite.app
+    rm -rf icons
+fi
 
 #
 # OpenInTerminal-Lite
